+++ common/src/main/java/excavated_variants/RegistryUtil.java
@@ -2,6 +2,7 @@ package excavated_variants;
 
 import dev.architectury.injectables.annotations.ExpectPlatform;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
 import org.jetbrains.annotations.Nullable;
 
@@ -9,6 +10,16 @@ public class RegistryUtil {
     @ExpectPlatform
     @Nullable
     public static Block getBlockById(ResourceLocation rl) {
+        throw new AssertionError();
+    }
+    @ExpectPlatform
+    @Nullable
+    public static Item getItemById(ResourceLocation rl) {
+        throw new AssertionError();
+    }
+    @ExpectPlatform
+    @Nullable
+    public static ResourceLocation getRlByBlock(Block block) {
         throw new AssertionError();
     }
 }