+++ common/src/main/java/excavated_variants/compat/rei/ExcavatedVariantsServerPlugin.java
@@ -1,0 +1,29 @@
+package excavated_variants.compat.rei;
+
+import excavated_variants.ExcavatedVariants;
+import excavated_variants.Pair;
+import excavated_variants.RegistryUtil;
+import excavated_variants.data.BaseOre;
+import excavated_variants.data.BaseStone;
+import excavated_variants.recipe.OreConversionRecipe;
+import me.shedaniel.rei.api.client.registry.display.DisplayRegistry;
+import me.shedaniel.rei.api.common.category.CategoryIdentifier;
+import me.shedaniel.rei.api.common.display.Display;
+import me.shedaniel.rei.api.common.display.DisplaySerializerRegistry;
+import me.shedaniel.rei.api.common.plugins.REIServerPlugin;
+import net.minecraft.core.NonNullList;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.crafting.CraftingRecipe;
+import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraft.world.item.crafting.ShapelessRecipe;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import java.util.Objects;
+
+public class ExcavatedVariantsServerPlugin implements REIServerPlugin {
+
+}