+++ fabric/src/main/java/excavated_variants/fabric/ExcavatedVariantsFabric.java
@@ -1,9 +1,11 @@
 package excavated_variants.fabric;
 
 import excavated_variants.ExcavatedVariants;
+import excavated_variants.worldgen.OreFinderUtil;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.biome.v1.BiomeModifications;
 import net.fabricmc.fabric.api.biome.v1.ModificationPhase;
+import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.minecraft.core.Registry;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
@@ -14,12 +16,17 @@ public class ExcavatedVariantsFabric implements ModInitializer {
     @Override
     public void onInitialize() {
         ExcavatedVariants.init();
+        ServerLifecycleEvents.SERVER_STARTING.register(server -> {
+            ExcavatedVariants.oreStoneList = null;
+            OreFinderUtil.reset();
+            ExcavatedVariants.setupMap();
+        });
         if (ExcavatedVariants.getConfig().attempt_ore_replacement) {
             ResourceKey<PlacedFeature> confKey = ResourceKey.create(Registry.PLACED_FEATURE_REGISTRY,new ResourceLocation(ExcavatedVariants.MOD_ID,"ore_replacer"));
             // A bit hacky, but will hopefully put it after existing stuff (like Unearthed's generation)
             BiomeModifications.create(confKey.location()).add(ModificationPhase.POST_PROCESSING, (x)->true, context -> {
                 context.getGenerationSettings().addFeature(GenerationStep.Decoration.TOP_LAYER_MODIFICATION, confKey);
-            });;
+            });
         }
     }
 }